Final changes to ForkSpec
